+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * 14 palettes have been added from the soilpaletttes package. (#95)
 * 20 palettes have been added from the rockthemes package. (#97)
+* 19 palettes have been added from the Manu package. (#98)
+
 
 # paletteer 1.2.0